remove network change detection propagation
