+++ libs/sdk-core/src/greenlight/node_api.rs
@@ -1,6 +1,5 @@
 use std::cmp::{min, Reverse};
 use std::collections::{HashMap, HashSet};
-use std::error::Error;
 use std::iter::Iterator;
 use std::pin::Pin;
 use std::str::FromStr;
@@ -53,6 +52,7 @@ use crate::lightning_invoice::{RawBolt11Invoice, SignedRawBolt11Invoice};
 use crate::node_api::{CreateInvoiceRequest, FetchBolt11Result, NodeAPI, NodeError, NodeResult};
 use crate::persist::db::SqliteStorage;
 use crate::persist::send_pays::{SendPay, SendPayStatus};
+use crate::tonic_wrap::with_connection_fallback;
 use crate::{models::*, LspInformation};
 use crate::{NodeConfig, PrepareRedeemOnchainFundsRequest, PrepareRedeemOnchainFundsResponse};
 
@@ -73,7 +73,6 @@ pub(crate) struct Greenlight {
     node_client: Mutex<Option<ClnClient>>,
     persister: Arc<SqliteStorage>,
     inprogress_payments: AtomicU16,
-    network_change_detector: watch::Sender<()>,
 }
 
 #[derive(Serialize, Deserialize)]
@@ -102,7 +101,6 @@ impl Greenlight {
         seed: Vec<u8>,
         restore_only: Option<bool>,
         persister: Arc<SqliteStorage>,
-        network_change_detector: watch::Sender<()>,
     ) -> NodeResult<Self> {
         // Derive the encryption key from the seed
         let temp_signer = Arc::new(Signer::new(
@@ -169,7 +167,6 @@ impl Greenlight {
                             seed,
                             creds.clone(),
                             persister,
-                            network_change_detector,
                         )
                     }
                     None => Err(NodeError::generic("Failed to encrypt credentials")),
@@ -184,7 +181,6 @@ impl Greenlight {
         seed: Vec<u8>,
         device: Device,
         persister: Arc<SqliteStorage>,
-        network_change_detector: watch::Sender<()>,
     ) -> NodeResult<Greenlight> {
         let greenlight_network = sdk_config.network.into();
         let signer = Signer::new(seed.clone(), greenlight_network, device.clone())?;
@@ -197,7 +193,6 @@ impl Greenlight {
             node_client: Mutex::new(None),
             persister,
             inprogress_payments: AtomicU16::new(0),
-            network_change_detector,
         })
     }
 
@@ -715,61 +710,6 @@ impl Greenlight {
         res
     }
 
-    async fn with_connection_fallback<T, M, F>(
-        &self,
-        main: M,
-        fallback: impl FnOnce() -> F,
-    ) -> Result<T, tonic::Status>
-    where
-        M: Future<Output = Result<T, tonic::Status>>,
-        F: Future<Output = Result<T, tonic::Status>>,
-        T: std::fmt::Debug,
-    {
-        let res = main.await;
-        let status = match res {
-            Ok(t) => return Ok(t),
-            Err(s) => s,
-        };
-        info!("AAAAAAAAAAAAAAA got status");
-
-        let source = match status.source() {
-            Some(source) => source,
-            None => return Err(status),
-        };
-        info!("AAAAAAAAAAAAAAA got source");
-
-        let error: &tonic::transport::Error = match source.downcast_ref() {
-            Some(error) => error,
-            None => return Err(status),
-        };
-        info!("AAAAAAAAAAAAAAA got error");
-
-        if error.to_string() != "transport error" {
-            return Err(status);
-        }
-        info!("AAAAAAAAAAAAAAA is transport error");
-
-        let source = match error.source() {
-            Some(source) => source,
-            None => return Err(status),
-        };
-        info!("AAAAAAAAAAAAAAA got source");
-        info!("AAAAAAAAAAAAAAA {}", source.to_string());
-        if !source.to_string().contains("keep-alive timed out") {
-            return Err(status);
-        }
-
-        info!("AAAAAAAAAAAAAAA is connection error");
-        if self.network_change_detector.send(()).is_err() {
-            warn!("greenlight network change detector died.");
-        }
-
-        let res = fallback().await;
-
-        info!("AAAAAAAAAAAAAAA fallback returned {:?}", res);
-        res
-    }
-
     // pulls transactions from greenlight based on last sync timestamp.
     // greenlight gives us the payments via API and for received payments we are looking for settled invoices.
     async fn pull_transactions(
@@ -1148,12 +1088,11 @@ impl NodeAPI for Greenlight {
             cltv: request.cltv,
         };
 
-        let res = self
-            .with_connection_fallback(client.invoice(cln_request.clone()), || {
-                client_clone.invoice(cln_request)
-            })
-            .await?
-            .into_inner();
+        let res = with_connection_fallback(client.invoice(cln_request.clone()), || {
+            client_clone.invoice(cln_request)
+        })
+        .await?
+        .into_inner();
         Ok(res.bolt11)
     }
 
@@ -1450,11 +1389,10 @@ impl NodeAPI for Greenlight {
             }),
         };
         let result: cln::PayResponse = self
-            .with_keep_alive(
-                self.with_connection_fallback(client.pay(request.clone()), || {
-                    client_clone.pay(request)
-                }),
-            )
+            .with_keep_alive(with_connection_fallback(
+                client.pay(request.clone()),
+                || client_clone.pay(request),
+            ))
             .await?
             .into_inner();
 
@@ -1495,11 +1433,10 @@ impl NodeAPI for Greenlight {
             maxfeepercent: fee_percent,
         };
         let result = self
-            .with_keep_alive(
-                self.with_connection_fallback(client.trampoline_pay(request.clone()), || {
-                    client_clone.trampoline_pay(request)
-                }),
-            )
+            .with_keep_alive(with_connection_fallback(
+                client.trampoline_pay(request.clone()),
+                || client_clone.trampoline_pay(request),
+            ))
             .await?
             .into_inner();
