+++ libs/sdk-core/src/tonic_wrap/mod.rs
@@ -1,4 +1,4 @@
-use std::{error::Error, fmt::Display};
+use std::{error::Error, fmt::Display, future::Future};
 
 pub struct Status(pub tonic::Status);
 
@@ -27,4 +27,54 @@ impl Display for TransportError {
             self.0.source(),
         )
     }
+}
+
+pub async fn with_connection_fallback<T, M, F>(
+    main: M,
+    fallback: impl FnOnce() -> F,
+) -> Result<T, tonic::Status>
+where
+    M: Future<Output = Result<T, tonic::Status>>,
+    F: Future<Output = Result<T, tonic::Status>>,
+    T: std::fmt::Debug,
+{
+    let res = main.await;
+    let status = match res {
+        Ok(t) => return Ok(t),
+        Err(s) => s,
+    };
+
+    debug!(
+        "with_connection_fallback: initial call failed with: {:?}",
+        status
+    );
+    let source = match status.source() {
+        Some(source) => source,
+        None => return Err(status),
+    };
+
+    let error: &tonic::transport::Error = match source.downcast_ref() {
+        Some(error) => error,
+        None => return Err(status),
+    };
+
+    if error.to_string() != "transport error" {
+        return Err(status);
+    }
+
+    let source = match error.source() {
+        Some(source) => source,
+        None => return Err(status),
+    };
+
+    if !source.to_string().contains("keep-alive timed out") {
+        return Err(status);
+    }
+
+    debug!(
+        "with_connection_fallback: initial call failed due to keepalive 
+        timeout. Retrying fallback."
+    );
+    let res = fallback().await;
+    res
 }
